EPMRPP-98732 || Refactor imports and update dependencies in build.gradle
