+++ build.gradle
@@ -32,13 +32,11 @@ ext['spring-boot.version'] = "${springBootVersion}"
 
 dependencies {
     if (releaseMode) {
-        implementation 'com.epam.reportportal:commons-dao'
-        implementation 'com.epam.reportportal:plugin-api'
-        annotationProcessor 'com.epam.reportportal:plugin-api'
+        implementation 'com.epam.reportportal:service-api'
+        annotationProcessor 'com.epam.reportportal:service-api'
     } else {
-        implementation 'com.github.reportportal:commons-dao:develop-SNAPSHOT'
-        implementation 'com.github.reportportal:plugin-api:develop-SNAPSHOT'
-        annotationProcessor 'com.github.reportportal:plugin-api:develop-SNAPSHOT'
+        implementation 'com.github.reportportal:service-api:9b01fc0'
+        annotationProcessor 'com.github.reportportal:service-api:9b01fc0'
     }
 
     // add lombok support
