+++ gradle.properties
@@ -2,5 +2,5 @@ version=5.13.1
 description=EPAM Report Portal. GitLab plugin.
 pluginId = GitLab
 
-springBootVersion=3.4.2
-lombokVersion=1.18.36
+springBootVersion=3.4.11
+lombokVersion=1.18.42
