+++ src/main/java/com/epam/reportportal/extension/gitlab/GitlabPlugin.java
@@ -9,8 +9,8 @@ import org.pf4j.PluginWrapper;
 public class GitlabPlugin extends Plugin {
 
   /**
-   * Constructor to be used by plugin manager for plugin instantiation. Your plugins have to provide
-   * constructor with this exact signature to be successfully loaded by manager.
+   * Constructor to be used by plugin manager for plugin instantiation. Your plugins have to provide constructor with
+   * this exact signature to be successfully loaded by manager.
    *
    * @param wrapper - A wrapper over plugin instance.
    */
