+++ src/main/java/com/epam/reportportal/extension/gitlab/client/GitlabClientProvider.java
@@ -1,9 +1,9 @@
 package com.epam.reportportal.extension.gitlab.client;
 
 import com.epam.reportportal.extension.gitlab.command.GitlabProperties;
-import com.epam.ta.reportportal.entity.integration.IntegrationParams;
-import com.epam.reportportal.rules.exception.ReportPortalException;
-import com.epam.reportportal.rules.exception.ErrorType;
+import com.epam.reportportal.infrastructure.persistence.entity.integration.IntegrationParams;
+import com.epam.reportportal.infrastructure.rules.exception.ErrorType;
+import com.epam.reportportal.infrastructure.rules.exception.ReportPortalException;
 import org.jasypt.util.text.BasicTextEncryptor;
 
 /**
