+++ src/main/java/com/epam/reportportal/extension/gitlab/command/GetIssueTypesCommand.java
@@ -16,9 +16,9 @@
 package com.epam.reportportal.extension.gitlab.command;
 
 import com.epam.reportportal.extension.ProjectMemberCommand;
-import com.epam.ta.reportportal.dao.ProjectRepository;
-import com.epam.ta.reportportal.dao.organization.OrganizationRepositoryCustom;
-import com.epam.ta.reportportal.entity.integration.Integration;
+import com.epam.reportportal.infrastructure.persistence.dao.ProjectRepository;
+import com.epam.reportportal.infrastructure.persistence.dao.organization.OrganizationRepositoryCustom;
+import com.epam.reportportal.infrastructure.persistence.entity.integration.Integration;
 import java.util.List;
 import java.util.Map;
 
@@ -30,7 +30,8 @@ public class GetIssueTypesCommand extends ProjectMemberCommand<List<String>> {
   public static final String ISSUE = "Issue";
   public static final String INCIDENT = "Incident";
 
-  public GetIssueTypesCommand(ProjectRepository projectRepository, OrganizationRepositoryCustom organizationRepository) {
+  public GetIssueTypesCommand(ProjectRepository projectRepository,
+      OrganizationRepositoryCustom organizationRepository) {
     super(projectRepository, organizationRepository);
   }
