+++ src/main/java/com/epam/reportportal/extension/gitlab/command/GitlabProperties.java
@@ -15,7 +15,7 @@
  */
 package com.epam.reportportal.extension.gitlab.command;
 
-import com.epam.ta.reportportal.entity.integration.IntegrationParams;
+import com.epam.reportportal.infrastructure.persistence.entity.integration.IntegrationParams;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
