+++ src/main/java/com/epam/reportportal/extension/gitlab/command/SearchEpicsCommand.java
@@ -21,10 +21,10 @@ import static java.util.Optional.ofNullable;
 import com.epam.reportportal.extension.PluginCommand;
 import com.epam.reportportal.extension.gitlab.client.GitlabClientProvider;
 import com.epam.reportportal.extension.gitlab.dto.EpicDto;
-import com.epam.reportportal.rules.exception.ErrorType;
-import com.epam.reportportal.rules.exception.ReportPortalException;
-import com.epam.ta.reportportal.entity.integration.Integration;
-import com.epam.ta.reportportal.entity.integration.IntegrationParams;
+import com.epam.reportportal.infrastructure.persistence.entity.integration.Integration;
+import com.epam.reportportal.infrastructure.persistence.entity.integration.IntegrationParams;
+import com.epam.reportportal.infrastructure.rules.exception.ErrorType;
+import com.epam.reportportal.infrastructure.rules.exception.ReportPortalException;
 import java.util.List;
 import java.util.Map;
 import org.slf4j.Logger;
@@ -34,6 +34,7 @@ import org.slf4j.LoggerFactory;
  * @author <a href="mailto:pavel_bortnik@epam.com">Pavel Bortnik</a>
  */
 public class SearchEpicsCommand implements PluginCommand<List<EpicDto>> {
+
   private static final Logger LOGGER = LoggerFactory.getLogger(SearchEpicsCommand.class);
 
   private final GitlabClientProvider gitlabClientProvider;
