+++ src/main/java/com/epam/reportportal/extension/gitlab/event/plugin/PluginEventListener.java
@@ -15,34 +15,35 @@
  */
 package com.epam.reportportal.extension.gitlab.event.plugin;
 
+import static java.util.Optional.ofNullable;
+
 import com.epam.reportportal.extension.event.PluginEvent;
 import com.epam.reportportal.extension.gitlab.event.EventHandlerFactory;
 import org.springframework.context.ApplicationListener;
-
-import static java.util.Optional.ofNullable;
 
 /**
  * @author <a href="mailto:ivan_budayeu@epam.com">Ivan Budayeu</a>
  */
 public class PluginEventListener implements ApplicationListener<PluginEvent> {
 
-    private final String pluginId;
-    private final EventHandlerFactory<PluginEvent> pluginEventEventHandlerFactory;
+  private final String pluginId;
+  private final EventHandlerFactory<PluginEvent> pluginEventEventHandlerFactory;
 
-    public PluginEventListener(String pluginId, EventHandlerFactory<PluginEvent> pluginEventEventHandlerFactory) {
-        this.pluginId = pluginId;
-        this.pluginEventEventHandlerFactory = pluginEventEventHandlerFactory;
-    }
+  public PluginEventListener(String pluginId, EventHandlerFactory<PluginEvent> pluginEventEventHandlerFactory) {
+    this.pluginId = pluginId;
+    this.pluginEventEventHandlerFactory = pluginEventEventHandlerFactory;
+  }
 
-    @Override
-    public void onApplicationEvent(PluginEvent event) {
-        if (supports(event)) {
-            ofNullable(pluginEventEventHandlerFactory.getEventHandler(event.getType())).ifPresent(pluginEventEventHandler -> pluginEventEventHandler
-                    .handle(event));
-        }
+  @Override
+  public void onApplicationEvent(PluginEvent event) {
+    if (supports(event)) {
+      ofNullable(pluginEventEventHandlerFactory.getEventHandler(event.getType())).ifPresent(
+          pluginEventEventHandler -> pluginEventEventHandler
+              .handle(event));
     }
+  }
 
-    private boolean supports(PluginEvent event) {
-        return pluginId.equals(event.getPluginId());
-    }
+  private boolean supports(PluginEvent event) {
+    return pluginId.equals(event.getPluginId());
+  }
 }
