+++ src/main/java/com/epam/reportportal/extension/gitlab/info/PluginInfoProvider.java
@@ -1,11 +1,11 @@
 package com.epam.reportportal.extension.gitlab.info;
 
-import com.epam.ta.reportportal.entity.integration.IntegrationType;
+import com.epam.reportportal.infrastructure.persistence.entity.integration.IntegrationType;
 
 /**
  * @author <a href="mailto:ivan_budayeu@epam.com">Ivan Budayeu</a>
  */
 public interface PluginInfoProvider {
 
-	IntegrationType provide(IntegrationType integrationType);
+  IntegrationType provide(IntegrationType integrationType);
 }
