+++ src/main/java/com/epam/reportportal/extension/gitlab/info/impl/PluginInfoProviderImpl.java
@@ -3,9 +3,9 @@ package com.epam.reportportal.extension.gitlab.info.impl;
 import static java.util.Optional.ofNullable;
 
 import com.epam.reportportal.extension.gitlab.info.PluginInfoProvider;
-import com.epam.ta.reportportal.entity.integration.IntegrationType;
-import com.epam.reportportal.rules.exception.ReportPortalException;
-import com.epam.reportportal.rules.exception.ErrorType;
+import com.epam.reportportal.infrastructure.persistence.entity.integration.IntegrationType;
+import com.epam.reportportal.infrastructure.rules.exception.ErrorType;
+import com.epam.reportportal.infrastructure.rules.exception.ReportPortalException;
 import java.io.IOException;
 import java.io.InputStream;
 import java.nio.file.Files;
