+++ src/main/java/com/epam/reportportal/extension/gitlab/utils/GitlabObjectMapperProvider.java
@@ -10,21 +10,22 @@ import com.fasterxml.jackson.databind.SerializationFeature;
  * @author Zsolt Nagyaghy
  */
 public class GitlabObjectMapperProvider {
-    private final ObjectMapper objectMapper;
 
-    public GitlabObjectMapperProvider() {
-        this.objectMapper = new ObjectMapper();
+  private final ObjectMapper objectMapper;
 
-        objectMapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
-        objectMapper.setPropertyNamingStrategy(PropertyNamingStrategy.SNAKE_CASE);
+  public GitlabObjectMapperProvider() {
+    this.objectMapper = new ObjectMapper();
 
-        objectMapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false);
-        objectMapper.configure(SerializationFeature.WRITE_ENUMS_USING_TO_STRING, true);
-        objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
-        objectMapper.configure(DeserializationFeature.READ_ENUMS_USING_TO_STRING, true);
-    }
+    objectMapper.setSerializationInclusion(JsonInclude.Include.NON_NULL);
+    objectMapper.setPropertyNamingStrategy(PropertyNamingStrategy.SNAKE_CASE);
 
-    public ObjectMapper getObjectMapper() {
-        return objectMapper;
-    }
+    objectMapper.configure(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS, false);
+    objectMapper.configure(SerializationFeature.WRITE_ENUMS_USING_TO_STRING, true);
+    objectMapper.configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false);
+    objectMapper.configure(DeserializationFeature.READ_ENUMS_USING_TO_STRING, true);
+  }
+
+  public ObjectMapper getObjectMapper() {
+    return objectMapper;
+  }
 }
