+++ src/main/java/com/epam/reportportal/extension/gitlab/utils/MemoizingSupplier.java
@@ -1,38 +1,38 @@
 package com.epam.reportportal.extension.gitlab.utils;
 
+import static com.google.common.base.Preconditions.checkNotNull;
+
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.function.Supplier;
-
-import static com.google.common.base.Preconditions.checkNotNull;
 
 /**
  * @author Andrei Piankouski
  */
 public class MemoizingSupplier<T> implements Supplier<T> {
 
-    private final Supplier<T> delegate;
+  private final Supplier<T> delegate;
 
-    private AtomicBoolean initialized = new AtomicBoolean(false);
+  private AtomicBoolean initialized = new AtomicBoolean(false);
 
-    private T value;
+  private T value;
 
-    public MemoizingSupplier(Supplier<T> delegate) {
-        this.delegate = checkNotNull(delegate);
-    }
+  public MemoizingSupplier(Supplier<T> delegate) {
+    this.delegate = checkNotNull(delegate);
+  }
 
-    @Override
-    public T get() {
+  @Override
+  public T get() {
+    if (!initialized.get()) {
+      synchronized (this) {
         if (!initialized.get()) {
-            synchronized (this) {
-                if (!initialized.get()) {
-                    T t = delegate.get();
-                    value = t;
-                    initialized.set(true);
-                    return t;
-                }
-            }
+          T t = delegate.get();
+          value = t;
+          initialized.set(true);
+          return t;
         }
-        return value;
+      }
     }
+    return value;
+  }
 
 }
